% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPhieRmix.R
\name{CRISPhieRmix}
\alias{CRISPhieRmix}
\alias{CRISPhieRmix}
\title{CRISPhieRmix}
\usage{
CRISPhieRmix(x, geneIds, negCtrl, max_iter = 100, tol = 1e-10, pq = 0.1,
  mu = 4, sigma = 1, nMesh = 100, maxDegree = 20, minDegree = 4,
  VERBOSE = FALSE, PLOT = FALSE)
}
\arguments{
\item{x}{log2 fold changes of guides targeting genes}

\item{geneIds}{gene ids corresponding to x}

\item{negCtrl}{log2 fold changes of negative control guides}

\item{max_iter}{maximum number of iterations for EM algorithm, default = 100}

\item{tol}{tolerance for convergence of EM algorithm, default = 1e-10}

\item{pq}{initial value of p*q, default = 0.1}

\item{mu}{initial value of mu for the interesting genes, default = 4}

\item{sigma}{initial value of sigma for the interesting genes, default = 1}

\item{nMesh}{the number of points to use in numerical integration of posterior probabilities, default = 100}

\item{maxDegree}{the maximum degree of the Lindsey fit, default = 20}

\item{minDegree}{the minimum degree of the Lindsey fit, default = 4}

\item{VERBOSE}{boolean variable for VERBOSE mode, default = FALSE}

\item{PLOT}{boolean variable to produce plots, default = FALSE}
}
\value{
a list containing genes, the corresponding posterior probabilities of being non-null, 
and the mixture fit

a list containing genes, the corresponding posterior probabilities, 
and the mixture fit.
}
\description{
A tool for indentifying interesting genes in large pooled CRISPRi and CRISPRa screen
using hierarchical mixture models

a hierarchical mixture model for analysing large-scale CRISPRi/a pooled screen
}
\examples{
Rosenbluh2017CRISPRi.sim.DESeq.log2fc.CRISPhieRmix = 
CRISPhieRmix(x = Rosenbluh2017CRISPRiSim$x, geneIds = Rosenbluh2017CRISPRiSim$geneIds,
negCtrl = Rosenbluh2017CRISPRiSim$negCtrl, mu = -2, sigma = 0.5, nMesh = 200)

}
\author{
Timothy Daley, \email{tdaley@stanford.edu}
}
