% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPhieRmix.R
\name{normalMix}
\alias{normalMix}
\title{normalMix}
\usage{
normalMix(x, geneIds, pq = 0.1, mu0 = 0, sigma0 = 1, mu = 5,
  sigma = 1, nMesh = 100, VERBOSE = FALSE, PLOT = FALSE)
}
\arguments{
\item{x}{log2fc of gene targetting guides}

\item{geneIds}{gene ids corresponding to x}

\item{pq}{initial value of p*q}

\item{mu0}{initial value of mean for the null distribution, default = 0
note that label switching can occur and this should be set with prior knowledge}

\item{sigma0}{initial value of sd for the null distribution, default = 1}

\item{mu}{initial value of the mean for the alternative distribution, default = 5}

\item{sigma}{initial value of the sd for the alternative distribution, default = 1}

\item{nMesh}{number of points for computing the gene posterior probabilities, default = 100}

\item{VERBOSE}{default = FALSE}

\item{PLOT}{default = FALSE}
}
\value{
a list containing genes, the corresponding posterior probabilities,
and the normal mixture fit.
}
\description{
fit a normal hierarchical 2 groups model to log2fc changes from a CRISPRi/a screen
}
\examples{
Rosenbluh2017CRISPRi.sim.DESeq.log2fc.NormalMix =
normalMix(x = Rosenbluh2017CRISPRiSim$x, geneIds = Rosenbluh2017CRISPRiSim$geneIds,
mu0 = 0, sigma0 = 0.5, mu = -2, sigma = 1, nMesh = 200)

}
